LangGraph-Research-Agent/requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-text-splitters
sentence-transformers
pypdf
docx2txt
unstructured
fastapi
uvicorn
streamlit
python-dotenv
pinecone 
langchain-pinecone 
python-multipart
langchain-groq
langchain-tavily
requests 
uuid 
langchain-huggingface
langchain-experimental
qdrant-client
langchain-google-genai
langchain_qdrant